cleaning up librispeech organization
