+++ algorithmic_efficiency/workloads/librispeech_conformer/workload.py
@@ -1,15 +1,7 @@
 import math
-from typing import Optional
 
-from absl import flags
-import torch
 
-from algorithmic_efficiency import data_utils
 from algorithmic_efficiency import spec
-from algorithmic_efficiency.workloads.librispeech_conformer.librispeech_pytorch.libri_dataset import \
-    LibriSpeechDataset
-
-FLAGS = flags.FLAGS
 
 
 class BaseLibrispeechWorkload(spec.Workload):
@@ -74,66 +66,6 @@ class BaseLibrispeechWorkload(spec.Workload):
   @property
   def eval_period_time_sec(self) -> int:
     return 40 * 60  # 40m
-
-  def _build_input_queue(self,
-                         data_rng: spec.RandomState,
-                         split: str,
-                         data_dir: str,
-                         global_batch_size: int,
-                         cache: Optional[bool] = False,
-                         repeat_final_dataset: Optional[bool] = False,
-                         num_batches: Optional[int] = None):
-    del cache
-    del repeat_final_dataset
-    return self._build_dataset(data_rng,
-                               split,
-                               data_dir,
-                               global_batch_size,
-                               num_batches)
-
-  def _build_dataset(self,
-                     data_rng: spec.RandomState,
-                     split: str,
-                     data_dir: str,
-                     global_batch_size: int,
-                     num_batches: Optional[int] = None):
-    train = False
-    if split == 'train':
-      split = 'train-clean-100+train-clean-360+train-other-500'
-      train = True
-    elif split == 'eval_train':
-      split = 'train-clean-100+train-clean-360+train-other-500'
-    elif split == 'validation':
-      split = 'dev-clean+dev-other'
-    elif split == 'test':
-      split = 'test-clean'
-
-    ds = LibriSpeechDataset(split=split, data_dir=data_dir)
-
-    dataloader = data_utils.cycle(
-        torch.utils.data.DataLoader(
-            ds,
-            batch_size=global_batch_size,
-            shuffle=train,
-            sampler=None,
-            num_workers=4,
-            prefetch_factor=10,
-            pin_memory=False,
-            drop_last=train,
-        ))
-
-    for batch in iter(dataloader):
-      inputs, input_paddings = batch['inputs']
-      targets, target_paddings = batch['targets']
-
-      numpy_batch = {
-          'inputs': (inputs.numpy(), input_paddings.numpy()),
-          'targets': (targets.numpy(), target_paddings.numpy()),
-      }
-
-      padded_batch = data_utils.shard_and_maybe_pad_np(
-          numpy_batch, padding_value=1.0, global_batch_size=global_batch_size)
-      yield padded_batch
 
   @property
   def step_hint(self) -> int:
